requirements.txt
nicegui>=1.4.20
binance-connector>=3.2.0
python-dotenv>=1.0.0
